Move old reading script tests into own file.
